Correccion ColisionBorder

+++ Codigopool/src/codigopool/BolaJugador.java
@@ -5,7 +5,7 @@ import java.awt.Graphics;
 
 public class BolaJugador extends Bola{
     
-    public BolaJugador(int x, int y, int n){
+    public BolaJugador(float x, float y, int n){
         super(x,y,n);
         
     }
